Refactor path require implementation
